+++ app/server/appsmith-server/src/main/java/com/appsmith/server/actioncollections/export/ActionCollectionExportableServiceCEImpl.java
@@ -9,7 +9,7 @@ import com.appsmith.server.dtos.ActionCollectionDTO;
 import com.appsmith.server.dtos.ApplicationJson;
 import com.appsmith.server.dtos.ExportingMetaDTO;
 import com.appsmith.server.dtos.MappedExportableResourcesDTO;
-import com.appsmith.server.export.exportable.ExportableServiceCE;
+import com.appsmith.server.exports.exportable.ExportableServiceCE;
 import com.appsmith.server.helpers.ImportExportUtils;
 import com.appsmith.server.solutions.ActionPermission;
 import reactor.core.publisher.Flux;
