+++ app/server/appsmith-server/src/main/java/com/appsmith/server/actioncollections/export/ActionCollectionExportableServiceImpl.java
@@ -2,7 +2,7 @@ package com.appsmith.server.actioncollections.export;
 
 import com.appsmith.server.actioncollections.base.ActionCollectionService;
 import com.appsmith.server.domains.ActionCollection;
-import com.appsmith.server.export.exportable.ExportableService;
+import com.appsmith.server.exports.exportable.ExportableService;
 import com.appsmith.server.solutions.ActionPermission;
 import org.springframework.stereotype.Service;
