+++ app/server/appsmith-server/src/main/java/com/appsmith/server/datasources/export/DatasourceExportableServiceImpl.java
@@ -3,7 +3,7 @@ package com.appsmith.server.datasources.export;
 import com.appsmith.external.models.Datasource;
 import com.appsmith.server.datasources.base.DatasourceService;
 import com.appsmith.server.datasourcestorages.base.DatasourceStorageService;
-import com.appsmith.server.export.exportable.ExportableService;
+import com.appsmith.server.exports.exportable.ExportableService;
 import com.appsmith.server.services.WorkspaceService;
 import com.appsmith.server.solutions.DatasourcePermission;
 import org.springframework.stereotype.Service;
