--- app/server/appsmith-server/src/main/java/com/appsmith/server/export/internal/ImportExportApplicationService.java
@@ -1,3 +1,0 @@
-package com.appsmith.server.export.internal;
-
-public interface ImportExportApplicationService extends ImportExportApplicationServiceCE {}
