+++ app/server/appsmith-server/src/main/java/com/appsmith/server/exports/internal/ExportApplicationService.java
@@ -1,0 +1,3 @@
+package com.appsmith.server.exports.internal;
+
+public interface ExportApplicationService extends ExportApplicationServiceCE {}
