+++ app/server/appsmith-server/src/main/java/com/appsmith/server/exports/internal/ExportApplicationServiceImpl.java
@@ -1,0 +1,56 @@
+package com.appsmith.server.exports.internal;
+
+import com.appsmith.external.models.Datasource;
+import com.appsmith.server.domains.ActionCollection;
+import com.appsmith.server.domains.CustomJSLib;
+import com.appsmith.server.domains.NewAction;
+import com.appsmith.server.domains.NewPage;
+import com.appsmith.server.domains.Plugin;
+import com.appsmith.server.domains.Theme;
+import com.appsmith.server.exports.exportable.ExportableService;
+import com.appsmith.server.exports.exportable.ExportableServiceCE;
+import com.appsmith.server.services.AnalyticsService;
+import com.appsmith.server.services.ApplicationService;
+import com.appsmith.server.services.SessionUserService;
+import com.appsmith.server.services.WorkspaceService;
+import com.appsmith.server.solutions.ApplicationPermission;
+import com.google.gson.Gson;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Primary;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+@Primary
+public class ExportApplicationServiceImpl extends ExportApplicationServiceCEImpl implements ExportApplicationService {
+
+    public ExportApplicationServiceImpl(
+            SessionUserService sessionUserService,
+            WorkspaceService workspaceService,
+            ApplicationService applicationService,
+            AnalyticsService analyticsService,
+            ApplicationPermission applicationPermission,
+            Gson gson,
+            ExportableService<Datasource> datasourceExportableService,
+            ExportableService<Plugin> pluginExportableService,
+            ExportableService<NewPage> newPageExportableService,
+            ExportableService<NewAction> newActionExportableService,
+            ExportableService<ActionCollection> actionCollectionExportableService,
+            ExportableServiceCE<Theme> themeExportableService,
+            ExportableService<CustomJSLib> customJSLibExportableService) {
+        super(
+                sessionUserService,
+                workspaceService,
+                applicationService,
+                analyticsService,
+                applicationPermission,
+                gson,
+                datasourceExportableService,
+                pluginExportableService,
+                newPageExportableService,
+                newActionExportableService,
+                actionCollectionExportableService,
+                themeExportableService,
+                customJSLibExportableService);
+    }
+}
