+++ app/server/appsmith-server/src/main/java/com/appsmith/server/fork/internal/ApplicationForkingServiceImpl.java
@@ -2,9 +2,9 @@ package com.appsmith.server.fork.internal;
 
 import com.appsmith.external.models.Datasource;
 import com.appsmith.server.actioncollections.base.ActionCollectionService;
-import com.appsmith.server.export.internal.ImportExportApplicationService;
 import com.appsmith.server.fork.forkable.ForkableService;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.imports.internal.ImportApplicationService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.repositories.ActionCollectionRepository;
 import com.appsmith.server.repositories.NewActionRepository;
@@ -38,7 +38,7 @@ public class ApplicationForkingServiceImpl extends ApplicationForkingServiceCEIm
             ResponseUtils responseUtils,
             WorkspacePermission workspacePermission,
             ApplicationPermission applicationPermission,
-            ImportExportApplicationService importExportApplicationService,
+            ImportApplicationService importApplicationService,
             ApplicationPageService applicationPageService,
             NewPageRepository newPageRepository,
             NewActionService newActionService,
@@ -60,7 +60,7 @@ public class ApplicationForkingServiceImpl extends ApplicationForkingServiceCEIm
                 responseUtils,
                 workspacePermission,
                 applicationPermission,
-                importExportApplicationService,
+                importApplicationService,
                 applicationPageService,
                 newPageRepository,
                 newActionService,
