+++ app/server/appsmith-server/src/main/java/com/appsmith/server/imports/importable/ImportableServiceCE.java
@@ -1,0 +1,5 @@
+package com.appsmith.server.imports.importable;
+
+import com.appsmith.external.models.BaseDomain;
+
+public interface ImportableServiceCE<T extends BaseDomain> {}
