+++ app/server/appsmith-server/src/main/java/com/appsmith/server/imports/internal/ImportApplicationService.java
@@ -1,0 +1,3 @@
+package com.appsmith.server.imports.internal;
+
+public interface ImportApplicationService extends ImportApplicationServiceCE {}
