+++ app/server/appsmith-server/src/main/java/com/appsmith/server/plugins/export/PluginExportableServiceImpl.java
@@ -1,7 +1,7 @@
 package com.appsmith.server.plugins.export;
 
 import com.appsmith.server.domains.Plugin;
-import com.appsmith.server.export.exportable.ExportableService;
+import com.appsmith.server.exports.exportable.ExportableService;
 import com.appsmith.server.plugins.base.PluginService;
 import com.appsmith.server.services.WorkspaceService;
 import org.springframework.stereotype.Service;
