+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ApplicationSnapshotServiceImpl.java
@@ -1,7 +1,8 @@
 package com.appsmith.server.services;
 
-import com.appsmith.server.export.internal.ImportExportApplicationService;
+import com.appsmith.server.exports.internal.ExportApplicationService;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.imports.internal.ImportApplicationService;
 import com.appsmith.server.repositories.ApplicationSnapshotRepository;
 import com.appsmith.server.services.ce.ApplicationSnapshotServiceCEImpl;
 import com.appsmith.server.solutions.ApplicationPermission;
@@ -17,14 +18,16 @@ public class ApplicationSnapshotServiceImpl extends ApplicationSnapshotServiceCE
     public ApplicationSnapshotServiceImpl(
             ApplicationSnapshotRepository applicationSnapshotRepository,
             ApplicationService applicationService,
-            ImportExportApplicationService importExportApplicationService,
+            ImportApplicationService importApplicationService,
+            ExportApplicationService exportApplicationService,
             ApplicationPermission applicationPermission,
             Gson gson,
             ResponseUtils responseUtils) {
         super(
                 applicationSnapshotRepository,
                 applicationService,
-                importExportApplicationService,
+                importApplicationService,
+                exportApplicationService,
                 applicationPermission,
                 gson,
                 responseUtils);
