+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ApplicationTemplateServiceImpl.java
@@ -1,8 +1,9 @@
 package com.appsmith.server.services;
 
 import com.appsmith.server.configurations.CloudServicesConfig;
-import com.appsmith.server.export.internal.ImportExportApplicationService;
+import com.appsmith.server.exports.internal.ExportApplicationService;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.imports.internal.ImportApplicationService;
 import com.appsmith.server.services.ce.ApplicationTemplateServiceCEImpl;
 import com.appsmith.server.solutions.ApplicationPermission;
 import com.appsmith.server.solutions.ReleaseNotesService;
@@ -13,10 +14,12 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class ApplicationTemplateServiceImpl extends ApplicationTemplateServiceCEImpl
         implements ApplicationTemplateService {
+
     public ApplicationTemplateServiceImpl(
             CloudServicesConfig cloudServicesConfig,
             ReleaseNotesService releaseNotesService,
-            ImportExportApplicationService importExportApplicationService,
+            ImportApplicationService importApplicationService,
+            ExportApplicationService exportApplicationService,
             AnalyticsService analyticsService,
             UserDataService userDataService,
             ApplicationService applicationService,
@@ -25,7 +28,8 @@ public class ApplicationTemplateServiceImpl extends ApplicationTemplateServiceCE
         super(
                 cloudServicesConfig,
                 releaseNotesService,
-                importExportApplicationService,
+                importApplicationService,
+                exportApplicationService,
                 analyticsService,
                 userDataService,
                 applicationService,
