+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/ApplicationSnapshotServiceCEImpl.java
@@ -7,8 +7,9 @@ import com.appsmith.server.domains.ApplicationSnapshot;
 import com.appsmith.server.dtos.ApplicationJson;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
-import com.appsmith.server.export.internal.ImportExportApplicationService;
+import com.appsmith.server.exports.internal.ExportApplicationService;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.imports.internal.ImportApplicationService;
 import com.appsmith.server.repositories.ApplicationSnapshotRepository;
 import com.appsmith.server.services.ApplicationService;
 import com.appsmith.server.solutions.ApplicationPermission;
@@ -27,7 +28,8 @@ import java.util.List;
 public class ApplicationSnapshotServiceCEImpl implements ApplicationSnapshotServiceCE {
     private final ApplicationSnapshotRepository applicationSnapshotRepository;
     private final ApplicationService applicationService;
-    private final ImportExportApplicationService importExportApplicationService;
+    private final ImportApplicationService importApplicationService;
+    private final ExportApplicationService exportApplicationService;
     private final ApplicationPermission applicationPermission;
     private final Gson gson;
     private final ResponseUtils responseUtils;
@@ -42,7 +44,7 @@ public class ApplicationSnapshotServiceCEImpl implements ApplicationSnapshotServ
                 exportApplicationById method check for MANAGE_PERMISSION if SerialiseApplicationObjective=SHARE.
                 */
                 .flatMap(branchedAppId -> Mono.zip(
-                        importExportApplicationService.exportApplicationById(
+                        exportApplicationService.exportApplicationById(
                                 branchedAppId, SerialiseApplicationObjective.VERSION_CONTROL),
                         Mono.just(branchedAppId)))
                 .flatMapMany(objects -> {
@@ -85,7 +87,7 @@ public class ApplicationSnapshotServiceCEImpl implements ApplicationSnapshotServ
                     String applicationJsonString = objects.getT1();
                     Application application = objects.getT2();
                     ApplicationJson applicationJson = gson.fromJson(applicationJsonString, ApplicationJson.class);
-                    return importExportApplicationService.restoreSnapshot(
+                    return importApplicationService.restoreSnapshot(
                             application.getWorkspaceId(), applicationJson, application.getId(), branchName);
                 })
                 .flatMap(application -> applicationSnapshotRepository
