+++ app/server/appsmith-server/src/main/java/com/appsmith/server/themes/export/ThemeExportableServiceImpl.java
@@ -1,7 +1,7 @@
 package com.appsmith.server.themes.export;
 
 import com.appsmith.server.domains.Theme;
-import com.appsmith.server.export.exportable.ExportableService;
+import com.appsmith.server.exports.exportable.ExportableService;
 import com.appsmith.server.themes.base.ThemeService;
 import org.springframework.stereotype.Service;
