+++ app/server/appsmith-server/src/test/java/com/appsmith/server/services/PageServiceTest.java
@@ -29,10 +29,11 @@ import com.appsmith.server.dtos.PageDTO;
 import com.appsmith.server.dtos.PageNameIdDTO;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
-import com.appsmith.server.export.internal.ImportExportApplicationService;
+import com.appsmith.server.exports.internal.ExportApplicationService;
 import com.appsmith.server.helpers.MockPluginExecutor;
 import com.appsmith.server.helpers.PluginExecutorHelper;
 import com.appsmith.server.helpers.TextUtils;
+import com.appsmith.server.imports.internal.ImportApplicationService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.repositories.PermissionGroupRepository;
@@ -125,7 +126,10 @@ public class PageServiceTest {
     LayoutCollectionService layoutCollectionService;
 
     @Autowired
-    ImportExportApplicationService importExportApplicationService;
+    ImportApplicationService importApplicationService;
+
+    @Autowired
+    ExportApplicationService exportApplicationService;
 
     @Autowired
     PermissionGroupRepository permissionGroupRepository;
@@ -175,11 +179,11 @@ public class PageServiceTest {
                     application.getGitApplicationMetadata().setDefaultApplicationId(application.getId());
                     return applicationService
                             .save(application)
-                            .zipWhen(application1 -> importExportApplicationService.exportApplicationById(
+                            .zipWhen(application1 -> exportApplicationService.exportApplicationById(
                                     application1.getId(), gitData.getBranchName()));
                 })
                 // Assign the branchName to all the resources connected to the application
-                .flatMap(tuple -> importExportApplicationService.importApplicationInWorkspaceFromGit(
+                .flatMap(tuple -> importApplicationService.importApplicationInWorkspaceFromGit(
                         workspaceId, tuple.getT2(), tuple.getT1().getId(), gitData.getBranchName()))
                 .block();
     }
