+++ app/server/appsmith-server/src/test/java/com/appsmith/server/solutions/CreateDBTablePageSolutionTests.java
@@ -28,9 +28,10 @@ import com.appsmith.server.dtos.CRUDPageResponseDTO;
 import com.appsmith.server.dtos.PageDTO;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
-import com.appsmith.server.export.internal.ImportExportApplicationService;
+import com.appsmith.server.exports.internal.ExportApplicationService;
 import com.appsmith.server.helpers.MockPluginExecutor;
 import com.appsmith.server.helpers.PluginExecutorHelper;
+import com.appsmith.server.imports.internal.ImportApplicationService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.repositories.PluginRepository;
@@ -89,42 +90,43 @@ public class CreateDBTablePageSolutionTests {
     private final String UPDATE_QUERY = "UpdateQuery";
     private final String INSERT_QUERY = "InsertQuery";
     private final Map<String, String> actionNameToBodyMap = Map.of(
-            "DeleteQuery", "DELETE FROM sampleTable\n" + "  WHERE \"id\" = {{data_table.triggeredRow.id}};",
+            "DeleteQuery",
+            "DELETE FROM sampleTable\n" + "  WHERE \"id\" = {{data_table.triggeredRow.id}};",
             "InsertQuery",
-                    "INSERT INTO sampleTable (\n" + "\t\"field1.something\", \n"
-                            + "\t\"field2\",\n"
-                            + "\t\"field3\", \n"
-                            + "\t\"field4\"\n"
-                            + ")\n"
-                            + "VALUES (\n"
-                            + "\t\t\t\t{{insert_form.formData.field1.something}}, \n"
-                            + "\t\t\t\t{{insert_form.formData.field2}}, \n"
-                            + "\t\t\t\t{{insert_form.formData.field3}}, \n"
-                            + "\t\t\t\t{{insert_form.formData.field4}}\n"
-                            + ");",
+            "INSERT INTO sampleTable (\n" + "\t\"field1.something\", \n"
+                    + "\t\"field2\",\n"
+                    + "\t\"field3\", \n"
+                    + "\t\"field4\"\n"
+                    + ")\n"
+                    + "VALUES (\n"
+                    + "\t\t\t\t{{insert_form.formData.field1.something}}, \n"
+                    + "\t\t\t\t{{insert_form.formData.field2}}, \n"
+                    + "\t\t\t\t{{insert_form.formData.field3}}, \n"
+                    + "\t\t\t\t{{insert_form.formData.field4}}\n"
+                    + ");",
             "SelectQuery",
-                    "SELECT * FROM sampleTable\n"
-                            + "WHERE \"field1.something\" like '%{{data_table.searchText || \"\"}}%'\n"
-                            + "ORDER BY \"{{data_table.sortOrder.column || 'id'}}\" {{data_table.sortOrder.order || 'ASC'}}\n"
-                            + "LIMIT {{data_table.pageSize}}"
-                            + "OFFSET {{(data_table.pageNo - 1) * data_table.pageSize}};",
+            "SELECT * FROM sampleTable\n"
+                    + "WHERE \"field1.something\" like '%{{data_table.searchText || \"\"}}%'\n"
+                    + "ORDER BY \"{{data_table.sortOrder.column || 'id'}}\" {{data_table.sortOrder.order || 'ASC'}}\n"
+                    + "LIMIT {{data_table.pageSize}}"
+                    + "OFFSET {{(data_table.pageNo - 1) * data_table.pageSize}};",
             "UpdateQuery",
-                    "UPDATE sampleTable SET\n"
-                            + "\t\t\"field1.something\" = '{{update_form.fieldState.field1.something.isVisible ? update_form.formData.field1.something : update_form.sourceData.field1.something}}',\n"
-                            + "    \"field2\" = '{{update_form.fieldState.field2.isVisible ? update_form.formData.field2 : update_form.sourceData.field2}}',\n"
-                            + "    \"field3\" = '{{update_form.fieldState.field3.isVisible ? update_form.formData.field3 : update_form.sourceData.field3}}',\n"
-                            + "\t\t\"field4\" = '{{update_form.fieldState.field4.isVisible ? update_form.formData.field4 : update_form.sourceData.field4}}'\n"
-                            + "  WHERE \"id\" = {{data_table.selectedRow.id}};",
+            "UPDATE sampleTable SET\n"
+                    + "\t\t\"field1.something\" = '{{update_form.fieldState.field1.something.isVisible ? update_form.formData.field1.something : update_form.sourceData.field1.something}}',\n"
+                    + "    \"field2\" = '{{update_form.fieldState.field2.isVisible ? update_form.formData.field2 : update_form.sourceData.field2}}',\n"
+                    + "    \"field3\" = '{{update_form.fieldState.field3.isVisible ? update_form.formData.field3 : update_form.sourceData.field3}}',\n"
+                    + "\t\t\"field4\" = '{{update_form.fieldState.field4.isVisible ? update_form.formData.field4 : update_form.sourceData.field4}}'\n"
+                    + "  WHERE \"id\" = {{data_table.selectedRow.id}};",
             "UpdateActionWithLessColumns",
-                    "UPDATE limitedColumnTable SET\n"
-                            + "\t\t\"field1.something\" = '{{update_form.fieldState.field1.something.isVisible ? update_form.formData.field1.something : update_form.sourceData.field1.something}}'\n"
-                            + "  WHERE \"id\" = {{data_table.selectedRow.id}};",
+            "UPDATE limitedColumnTable SET\n"
+                    + "\t\t\"field1.something\" = '{{update_form.fieldState.field1.something.isVisible ? update_form.formData.field1.something : update_form.sourceData.field1.something}}'\n"
+                    + "  WHERE \"id\" = {{data_table.selectedRow.id}};",
             "InsertActionWithLessColumns",
-                    "INSERT INTO limitedColumnTable (\n" + "\t\"field1.something\" \n"
-                            + ")\n"
-                            + "VALUES (\n"
-                            + "\t\t\t\t{{insert_form.formData.field1.something}} \n"
-                            + ");");
+            "INSERT INTO limitedColumnTable (\n" + "\t\"field1.something\" \n"
+                    + ")\n"
+                    + "VALUES (\n"
+                    + "\t\t\t\t{{insert_form.formData.field1.something}} \n"
+                    + ");");
 
     @Autowired
     CreateDBTablePageSolution solution;
@@ -151,7 +153,10 @@ public class CreateDBTablePageSolutionTests {
     PluginRepository pluginRepository;
 
     @Autowired
-    ImportExportApplicationService importExportApplicationService;
+    ImportApplicationService importApplicationService;
+
+    @Autowired
+    ExportApplicationService exportApplicationService;
 
     @Autowired
     ApplicationService applicationService;
@@ -364,11 +369,11 @@ public class CreateDBTablePageSolutionTests {
                     gitData.setDefaultApplicationId(application.getId());
                     return applicationService
                             .save(application)
-                            .zipWhen(application1 -> importExportApplicationService.exportApplicationById(
+                            .zipWhen(application1 -> exportApplicationService.exportApplicationById(
                                     application1.getId(), gitData.getBranchName()));
                 })
                 // Assign the branchName to all the resources connected to the application
-                .flatMap(tuple -> importExportApplicationService.importApplicationInWorkspaceFromGit(
+                .flatMap(tuple -> importApplicationService.importApplicationInWorkspaceFromGit(
                         testWorkspace.getId(), tuple.getT2(), tuple.getT1().getId(), gitData.getBranchName()))
                 .block();
