+++ app/server/appsmith-server/src/test/java/com/appsmith/server/solutions/ImportApplicationServiceTests.java
@@ -43,9 +43,10 @@ import com.appsmith.server.dtos.PageDTO;
 import com.appsmith.server.dtos.PageNameIdDTO;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
-import com.appsmith.server.export.internal.ImportExportApplicationService;
+import com.appsmith.server.exports.internal.ExportApplicationService;
 import com.appsmith.server.helpers.MockPluginExecutor;
 import com.appsmith.server.helpers.PluginExecutorHelper;
+import com.appsmith.server.imports.internal.ImportApplicationService;
 import com.appsmith.server.jslibs.base.CustomJSLibService;
 import com.appsmith.server.migrations.ApplicationVersion;
 import com.appsmith.server.migrations.JsonSchemaMigration;
@@ -131,7 +132,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 @SpringBootTest
 @DirtiesContext
 @TestMethodOrder(MethodOrderer.MethodName.class)
-public class ImportExportApplicationServiceTests {
+public class ImportApplicationServiceTests {
 
     private static final String INVALID_JSON_FILE = "invalid json file";
     private static final Map<String, Datasource> datasourceMap = new HashMap<>();
@@ -145,7 +146,10 @@ public class ImportExportApplicationServiceTests {
     private static String exportWithConfigurationAppId;
 
     @Autowired
-    ImportExportApplicationService importExportApplicationService;
+    ExportApplicationService exportApplicationService;
+
+    @Autowired
+    ImportApplicationService importApplicationService;
 
     @Autowired
     Gson gson;
@@ -331,7 +335,7 @@ public class ImportExportApplicationServiceTests {
     @Test
     @WithUserDetails(value = "api_user")
     public void exportApplicationWithNullApplicationIdTest() {
-        Mono<ApplicationJson> resultMono = importExportApplicationService.exportApplicationById(null, "");
+        Mono<ApplicationJson> resultMono = exportApplicationService.exportApplicationById(null, "");
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -363,7 +367,7 @@ public class ImportExportApplicationServiceTests {
                 .block();
 
         Mono<ApplicationJson> resultMono =
-                importExportApplicationService.exportApplicationById(createdApplication.getId(), "");
+                exportApplicationService.exportApplicationById(createdApplication.getId(), "");
 
         StepVerifier.create(resultMono)
                 .assertNext(applicationJson -> {
@@ -379,7 +383,7 @@ public class ImportExportApplicationServiceTests {
     @Test
     @WithUserDetails(value = "api_user")
     public void exportApplication_withInvalidApplicationId_throwNoResourceFoundException() {
-        Mono<ApplicationJson> resultMono = importExportApplicationService.exportApplicationById("invalidAppId", "");
+        Mono<ApplicationJson> resultMono = exportApplicationService.exportApplicationById("invalidAppId", "");
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -393,7 +397,7 @@ public class ImportExportApplicationServiceTests {
     @Test
     @WithUserDetails(value = "api_user")
     public void exportApplicationById_WhenContainsInternalFields_InternalFieldsNotExported() {
-        Mono<ApplicationJson> resultMono = importExportApplicationService.exportApplicationById(testAppId, "");
+        Mono<ApplicationJson> resultMono = exportApplicationService.exportApplicationById(testAppId, "");
 
         StepVerifier.create(resultMono)
                 .assertNext(applicationJson -> {
@@ -430,7 +434,7 @@ public class ImportExportApplicationServiceTests {
 
                     return applicationPageService.createApplication(testApplication, workspaceId);
                 })
-                .flatMap(application -> importExportApplicationService.exportApplicationById(application.getId(), ""));
+                .flatMap(application -> exportApplicationService.exportApplicationById(application.getId(), ""));
 
         StepVerifier.create(resultMono)
                 .assertNext(applicationJson -> {
@@ -538,7 +542,7 @@ public class ImportExportApplicationServiceTests {
                             .then(layoutActionService.createSingleAction(action2, Boolean.FALSE))
                             .then(layoutActionService.updateLayout(
                                     testPage.getId(), testPage.getApplicationId(), layout.getId(), layout))
-                            .then(importExportApplicationService.exportApplicationById(testApp.getId(), ""));
+                            .then(exportApplicationService.exportApplicationById(testApp.getId(), ""));
                 })
                 .cache();
 
@@ -747,7 +751,7 @@ public class ImportExportApplicationServiceTests {
 
                     return layoutActionService
                             .createAction(action)
-                            .then(importExportApplicationService.exportApplicationById(
+                            .then(exportApplicationService.exportApplicationById(
                                     testApp.getId(), SerialiseApplicationObjective.VERSION_CONTROL));
                 });
 
@@ -819,7 +823,7 @@ public class ImportExportApplicationServiceTests {
         Mockito.when(filepart.headers().getContentType()).thenReturn(MediaType.IMAGE_PNG);
 
         Mono<ApplicationImportDTO> resultMono =
-                importExportApplicationService.extractFileAndSaveApplication(workspaceId, filepart);
+                importApplicationService.extractFileAndSaveApplication(workspaceId, filepart);
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(error -> error instanceof AppsmithException)
@@ -831,8 +835,7 @@ public class ImportExportApplicationServiceTests {
     public void importApplicationWithNullWorkspaceIdTest() {
         FilePart filepart = Mockito.mock(FilePart.class, Mockito.RETURNS_DEEP_STUBS);
 
-        Mono<ApplicationImportDTO> resultMono =
-                importExportApplicationService.extractFileAndSaveApplication(null, filepart);
+        Mono<ApplicationImportDTO> resultMono = importApplicationService.extractFileAndSaveApplication(null, filepart);
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -848,7 +851,7 @@ public class ImportExportApplicationServiceTests {
 
         FilePart filePart = createFilePart("test_assets/ImportExportServiceTest/invalid-json-without-pages.json");
         Mono<ApplicationImportDTO> resultMono =
-                importExportApplicationService.extractFileAndSaveApplication(workspaceId, filePart);
+                importApplicationService.extractFileAndSaveApplication(workspaceId, filePart);
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -865,7 +868,7 @@ public class ImportExportApplicationServiceTests {
 
         FilePart filePart = createFilePart("test_assets/ImportExportServiceTest/invalid-json-without-app.json");
         Mono<ApplicationImportDTO> resultMono =
-                importExportApplicationService.extractFileAndSaveApplication(workspaceId, filePart);
+                importApplicationService.extractFileAndSaveApplication(workspaceId, filePart);
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -893,7 +896,7 @@ public class ImportExportApplicationServiceTests {
                 .block();
 
         final Mono<ApplicationImportDTO> resultMono = workspaceMono.flatMap(
-                workspace -> importExportApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
+                workspace -> importApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
 
         List<PermissionGroup> permissionGroups = workspaceMono
                 .flatMapMany(savedWorkspace -> {
@@ -1058,7 +1061,7 @@ public class ImportExportApplicationServiceTests {
         newWorkspace.setName("Midway cancel import app workspace");
         newWorkspace = workspaceService.create(newWorkspace).block();
 
-        importExportApplicationService
+        importApplicationService
                 .extractFileAndSaveApplication(newWorkspace.getId(), filePart)
                 .timeout(Duration.ofMillis(10))
                 .subscribe();
@@ -1095,7 +1098,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<ApplicationImportDTO> resultMono = workspaceService
                 .create(newWorkspace)
                 .flatMap(workspace ->
-                        importExportApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
+                        importApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
 
         StepVerifier.create(resultMono.flatMap(applicationImportDTO -> Mono.zip(
                         Mono.just(applicationImportDTO),
@@ -1134,7 +1137,7 @@ public class ImportExportApplicationServiceTests {
         Mono<Workspace> workspaceMono = workspaceService.create(newWorkspace).cache();
 
         final Mono<ApplicationImportDTO> resultMono = workspaceMono.flatMap(
-                workspace -> importExportApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
+                workspace -> importApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
 
         List<PermissionGroup> permissionGroups = workspaceMono
                 .flatMapMany(savedWorkspace -> {
@@ -1251,7 +1254,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<ApplicationImportDTO> resultMono = workspaceService
                 .create(newWorkspace)
                 .flatMap(workspace ->
-                        importExportApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
+                        importApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
 
         StepVerifier.create(resultMono)
                 .assertNext(applicationImportDTO -> {
@@ -1276,7 +1279,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<ApplicationImportDTO> resultMono = workspaceService
                 .create(newWorkspace)
                 .flatMap(workspace ->
-                        importExportApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
+                        importApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
 
         StepVerifier.create(resultMono.flatMap(applicationImportDTO -> Mono.zip(
                         Mono.just(applicationImportDTO),
@@ -1355,9 +1358,9 @@ public class ImportExportApplicationServiceTests {
                             .createAction(action)
                             .flatMap(createdAction -> newActionService.findById(createdAction.getId(), READ_ACTIONS));
                 })
-                .then(importExportApplicationService
+                .then(exportApplicationService
                         .exportApplicationById(savedApplication.getId(), SerialiseApplicationObjective.VERSION_CONTROL)
-                        .flatMap(applicationJson -> importExportApplicationService.importApplicationInWorkspaceFromGit(
+                        .flatMap(applicationJson -> importApplicationService.importApplicationInWorkspaceFromGit(
                                 workspaceId, applicationJson, savedApplication.getId(), gitData.getBranchName())))
                 .cache();
 
@@ -1395,7 +1398,7 @@ public class ImportExportApplicationServiceTests {
     public void importApplication_incompatibleJsonFile_throwException() {
         FilePart filePart = createFilePart("test_assets/ImportExportServiceTest/incompatible_version.json");
         Mono<ApplicationImportDTO> resultMono =
-                importExportApplicationService.extractFileAndSaveApplication(workspaceId, filePart);
+                importApplicationService.extractFileAndSaveApplication(workspaceId, filePart);
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -1415,7 +1418,7 @@ public class ImportExportApplicationServiceTests {
         Mono<Workspace> workspaceMono = workspaceService.create(newWorkspace).cache();
 
         final Mono<ApplicationImportDTO> resultMono = workspaceMono.flatMap(
-                workspace -> importExportApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
+                workspace -> importApplicationService.extractFileAndSaveApplication(workspace.getId(), filePart));
 
         List<PermissionGroup> permissionGroups = workspaceMono
                 .flatMapMany(savedWorkspace -> {
@@ -1571,7 +1574,7 @@ public class ImportExportApplicationServiceTests {
                 .block();
         applicationJson.getPageList().get(0).setGitSyncId(gitSyncIdBeforeImport);
 
-        Application importedApplication = importExportApplicationService
+        Application importedApplication = importApplicationService
                 .importApplicationInWorkspaceFromGit(workspaceId, applicationJson, application.getId(), "master")
                 .block();
 
@@ -1640,7 +1643,7 @@ public class ImportExportApplicationServiceTests {
                 .block();
         applicationJson.getPageList().get(0).setGitSyncId(gitSyncIdBeforeImport);
 
-        Application applicationMono = importExportApplicationService
+        Application applicationMono = importApplicationService
                 .importApplicationInWorkspaceFromGit(workspaceId, applicationJson, application.getId(), "master")
                 .block();
 
@@ -1738,9 +1741,9 @@ public class ImportExportApplicationServiceTests {
                         anonymousPermissionGroup.getId()))
                 .build();
 
-        Mono<Application> applicationMono = importExportApplicationService
+        Mono<Application> applicationMono = exportApplicationService
                 .exportApplicationById(application.getId(), "master")
-                .flatMap(applicationJson -> importExportApplicationService.importApplicationInWorkspaceFromGit(
+                .flatMap(applicationJson -> importApplicationService.importApplicationInWorkspaceFromGit(
                         workspaceId, applicationJson, application.getId(), "master"));
 
         StepVerifier.create(applicationMono)
@@ -1774,7 +1777,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<Application> resultMonoWithoutDiscardOperation = applicationJsonMono
                 .flatMap(applicationJson -> {
                     applicationJson.getExportedApplication().setName("discard-change-page-added");
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspaceId, applicationJson);
                 })
                 .flatMap(application -> {
@@ -1841,7 +1844,7 @@ public class ImportExportApplicationServiceTests {
                             .setDefaultApplicationId(importedApplication.getId());
                     return applicationService
                             .save(importedApplication)
-                            .then(importExportApplicationService.importApplicationInWorkspaceFromGit(
+                            .then(importApplicationService.importApplicationInWorkspaceFromGit(
                                     importedApplication.getWorkspaceId(),
                                     applicationJson,
                                     importedApplication.getId(),
@@ -1894,7 +1897,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<Application> resultMonoWithoutDiscardOperation = applicationJsonMono
                 .flatMap(applicationJson -> {
                     applicationJson.getExportedApplication().setName("discard-change-action-added");
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspaceId, applicationJson);
                 })
                 .flatMap(application -> {
@@ -1940,7 +1943,7 @@ public class ImportExportApplicationServiceTests {
                             .setDefaultApplicationId(importedApplication.getId());
                     return applicationService
                             .save(importedApplication)
-                            .then(importExportApplicationService.importApplicationInWorkspaceFromGit(
+                            .then(importApplicationService.importApplicationInWorkspaceFromGit(
                                     importedApplication.getWorkspaceId(),
                                     applicationJson,
                                     importedApplication.getId(),
@@ -1987,7 +1990,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<Application> resultMonoWithoutDiscardOperation = applicationJsonMono
                 .flatMap(applicationJson -> {
                     applicationJson.getExportedApplication().setName("discard-change-collection-added");
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspaceId, applicationJson);
                 })
                 .flatMap(application -> {
@@ -2053,7 +2056,7 @@ public class ImportExportApplicationServiceTests {
                             .setDefaultApplicationId(importedApplication.getId());
                     return applicationService
                             .save(importedApplication)
-                            .then(importExportApplicationService.importApplicationInWorkspaceFromGit(
+                            .then(importApplicationService.importApplicationInWorkspaceFromGit(
                                     importedApplication.getWorkspaceId(),
                                     applicationJson,
                                     importedApplication.getId(),
@@ -2117,7 +2120,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<Application> resultMonoWithoutDiscardOperation = applicationJsonMono
                 .flatMap(applicationJson -> {
                     applicationJson.getExportedApplication().setName("discard-change-page-removed");
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspaceId, applicationJson);
                 })
                 .flatMap(application -> {
@@ -2156,7 +2159,7 @@ public class ImportExportApplicationServiceTests {
                             .setDefaultApplicationId(importedApplication.getId());
                     return applicationService
                             .save(importedApplication)
-                            .then(importExportApplicationService.importApplicationInWorkspaceFromGit(
+                            .then(importApplicationService.importApplicationInWorkspaceFromGit(
                                     importedApplication.getWorkspaceId(),
                                     applicationJson,
                                     importedApplication.getId(),
@@ -2198,7 +2201,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<Application> resultMonoWithoutDiscardOperation = applicationJsonMono
                 .flatMap(applicationJson -> {
                     applicationJson.getExportedApplication().setName("discard-change-action-removed");
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspaceId, applicationJson);
                 })
                 .flatMap(application -> {
@@ -2237,7 +2240,7 @@ public class ImportExportApplicationServiceTests {
                             .setDefaultApplicationId(importedApplication.getId());
                     return applicationService
                             .save(importedApplication)
-                            .then(importExportApplicationService.importApplicationInWorkspaceFromGit(
+                            .then(importApplicationService.importApplicationInWorkspaceFromGit(
                                     importedApplication.getWorkspaceId(),
                                     applicationJson,
                                     importedApplication.getId(),
@@ -2278,7 +2281,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<Application> resultMonoWithoutDiscardOperation = applicationJsonMono
                 .flatMap(applicationJson -> {
                     applicationJson.getExportedApplication().setName("discard-change-collection-removed");
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspaceId, applicationJson);
                 })
                 .flatMap(application -> {
@@ -2324,7 +2327,7 @@ public class ImportExportApplicationServiceTests {
                             .setDefaultApplicationId(importedApplication.getId());
                     return applicationService
                             .save(importedApplication)
-                            .then(importExportApplicationService.importApplicationInWorkspaceFromGit(
+                            .then(importApplicationService.importApplicationInWorkspaceFromGit(
                                     importedApplication.getWorkspaceId(),
                                     applicationJson,
                                     importedApplication.getId(),
@@ -2366,7 +2369,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<Application> resultMonoWithoutDiscardOperation = applicationJsonMono
                 .flatMap(applicationJson -> {
                     applicationJson.getExportedApplication().setName("discard-change-navsettings-added");
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspaceId, applicationJson);
                 })
                 .flatMap(application -> {
@@ -2415,7 +2418,7 @@ public class ImportExportApplicationServiceTests {
                             .setDefaultApplicationId(importedApplication.getId());
                     return applicationService
                             .save(importedApplication)
-                            .then(importExportApplicationService.importApplicationInWorkspaceFromGit(
+                            .then(importApplicationService.importApplicationInWorkspaceFromGit(
                                     importedApplication.getWorkspaceId(),
                                     applicationJson,
                                     importedApplication.getId(),
@@ -2448,7 +2451,7 @@ public class ImportExportApplicationServiceTests {
         final Mono<Application> resultMonoWithoutDiscardOperation = applicationJsonMono
                 .flatMap(applicationJson -> {
                     applicationJson.getExportedApplication().setName("discard-change-applayout-added");
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspaceId, applicationJson);
                 })
                 .flatMap(application -> {
@@ -2478,7 +2481,7 @@ public class ImportExportApplicationServiceTests {
                             .setDefaultApplicationId(importedApplication.getId());
                     return applicationService
                             .save(importedApplication)
-                            .then(importExportApplicationService.importApplicationInWorkspaceFromGit(
+                            .then(importApplicationService.importApplicationInWorkspaceFromGit(
                                     importedApplication.getWorkspaceId(),
                                     applicationJson,
                                     importedApplication.getId(),
@@ -2627,7 +2630,7 @@ public class ImportExportApplicationServiceTests {
                             .then(layoutActionService.createSingleAction(action2, Boolean.FALSE))
                             .then(layoutActionService.updateLayout(
                                     testPage.getId(), testPage.getApplicationId(), layout.getId(), layout))
-                            .then(importExportApplicationService.exportApplicationById(testApp.getId(), ""));
+                            .then(exportApplicationService.exportApplicationById(testApp.getId(), ""));
                 })
                 .cache();
 
@@ -2807,7 +2810,7 @@ public class ImportExportApplicationServiceTests {
     @Test
     @WithUserDetails(value = "usertest@usertest.com")
     public void exportApplication_withReadOnlyAccess_exportedWithDecryptedFields() {
-        Mono<ApplicationJson> exportApplicationMono = importExportApplicationService.exportApplicationById(
+        Mono<ApplicationJson> exportApplicationMono = exportApplicationService.exportApplicationById(
                 exportWithConfigurationAppId, SerialiseApplicationObjective.SHARE);
 
         StepVerifier.create(exportApplicationMono)
@@ -2847,7 +2850,7 @@ public class ImportExportApplicationServiceTests {
 
         datasourceService.create(testDatasource).block();
 
-        final Mono<Application> resultMono = importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+        final Mono<Application> resultMono = importApplicationService.importNewApplicationInWorkspaceFromJson(
                 testWorkspace.getId(), applicationJson);
 
         StepVerifier.create(resultMono.flatMap(application -> Mono.zip(
@@ -2910,7 +2913,7 @@ public class ImportExportApplicationServiceTests {
         testDatasource.setDatasourceStorages(storages);
         datasourceService.create(testDatasource).block();
 
-        final Mono<Application> resultMono = importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+        final Mono<Application> resultMono = importApplicationService.importNewApplicationInWorkspaceFromJson(
                 testWorkspace.getId(), applicationJson);
 
         StepVerifier.create(resultMono.flatMap(application -> Mono.zip(
@@ -2984,7 +2987,7 @@ public class ImportExportApplicationServiceTests {
         // Deploy the current application
         applicationPageService.publish(testApplication.getId(), true).block();
 
-        Mono<ApplicationJson> applicationJsonMono = importExportApplicationService
+        Mono<ApplicationJson> applicationJsonMono = exportApplicationService
                 .exportApplicationById(testApplication.getId(), "")
                 .cache();
 
@@ -3012,7 +3015,7 @@ public class ImportExportApplicationServiceTests {
                 .verifyComplete();
 
         ApplicationJson applicationJson = applicationJsonMono.block();
-        Application application = importExportApplicationService
+        Application application = importApplicationService
                 .importNewApplicationInWorkspaceFromJson(workspaceId, applicationJson)
                 .block();
 
@@ -3091,12 +3094,12 @@ public class ImportExportApplicationServiceTests {
                 })
                 .block();
 
-        Mono<Application> result = importExportApplicationService
+        Mono<Application> result = exportApplicationService
                 .exportApplicationById(savedApplication.getId(), SerialiseApplicationObjective.VERSION_CONTROL)
                 .flatMap(applicationJson -> {
                     // setting published mode resource as null, similar to the app json exported to git repo
                     applicationJson.getExportedApplication().setPublishedApplicationDetail(null);
-                    return importExportApplicationService.importApplicationInWorkspaceFromGit(
+                    return importApplicationService.importApplicationInWorkspaceFromGit(
                             workspaceId, applicationJson, savedApplication.getId(), gitData.getBranchName());
                 });
 
@@ -3146,12 +3149,12 @@ public class ImportExportApplicationServiceTests {
                 })
                 .block();
 
-        Mono<Application> result = importExportApplicationService
+        Mono<Application> result = exportApplicationService
                 .exportApplicationById(savedApplication.getId(), SerialiseApplicationObjective.VERSION_CONTROL)
                 .flatMap(applicationJson -> {
                     // setting published mode resource as null, similar to the app json exported to git repo
                     applicationJson.getExportedApplication().setPublishedAppLayout(null);
-                    return importExportApplicationService.importApplicationInWorkspaceFromGit(
+                    return importApplicationService.importApplicationInWorkspaceFromGit(
                             workspaceId, applicationJson, savedApplication.getId(), gitData.getBranchName());
                 });
 
@@ -3204,7 +3207,7 @@ public class ImportExportApplicationServiceTests {
                 .reorderPage(testApplication.getId(), testPage2.getId(), 1, null)
                 .block();
 
-        Mono<ApplicationJson> applicationJsonMono = importExportApplicationService
+        Mono<ApplicationJson> applicationJsonMono = exportApplicationService
                 .exportApplicationById(testApplication.getId(), "")
                 .cache();
 
@@ -3230,7 +3233,7 @@ public class ImportExportApplicationServiceTests {
                 .verifyComplete();
 
         ApplicationJson applicationJson = applicationJsonMono.block();
-        Application application = importExportApplicationService
+        Application application = importApplicationService
                 .importNewApplicationInWorkspaceFromJson(workspaceId, applicationJson)
                 .block();
 
@@ -3368,7 +3371,7 @@ public class ImportExportApplicationServiceTests {
         Mono<Tuple3<ApplicationPagesDTO, List<NewAction>, List<ActionCollection>>> tuple2Mono = createAppAndPageMono
                 .flatMap(application ->
                         // merge the application json with the application we've created
-                        importExportApplicationService
+                        importApplicationService
                                 .mergeApplicationJsonWithApplication(
                                         application.getWorkspaceId(), application.getId(), null, applicationJson, null)
                                 .thenReturn(application))
@@ -3432,7 +3435,7 @@ public class ImportExportApplicationServiceTests {
         Mono<ApplicationPagesDTO> applicationPagesDTOMono = createAppAndPageMono
                 .flatMap(application ->
                         // merge the application json with the application we've created
-                        importExportApplicationService
+                        importApplicationService
                                 .mergeApplicationJsonWithApplication(
                                         application.getWorkspaceId(),
                                         application.getId(),
@@ -3473,8 +3476,7 @@ public class ImportExportApplicationServiceTests {
                     String branchName = null;
                     return applicationService
                             .save(application)
-                            .then(importExportApplicationService.exportApplicationById(
-                                    application.getId(), branchName));
+                            .then(exportApplicationService.exportApplicationById(application.getId(), branchName));
                 });
 
         StepVerifier.create(exportedAppJson)
@@ -3500,8 +3502,8 @@ public class ImportExportApplicationServiceTests {
         final String randomId = UUID.randomUUID().toString();
         appJson.getDatasourceList().get(0).setPluginId(randomId);
         Workspace createdWorkspace = workspaceService.create(newWorkspace).block();
-        final Mono<Application> resultMono = importExportApplicationService.importNewApplicationInWorkspaceFromJson(
-                createdWorkspace.getId(), appJson);
+        final Mono<Application> resultMono =
+                importApplicationService.importNewApplicationInWorkspaceFromJson(createdWorkspace.getId(), appJson);
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -3529,7 +3531,7 @@ public class ImportExportApplicationServiceTests {
                 .flatMap(tuple -> {
                     Workspace workspace = tuple.getT1();
                     ApplicationJson applicationJson = tuple.getT2();
-                    return importExportApplicationService.importNewApplicationInWorkspaceFromJson(
+                    return importApplicationService.importNewApplicationInWorkspaceFromJson(
                             workspace.getId(), applicationJson);
                 });
 
@@ -3562,7 +3564,7 @@ public class ImportExportApplicationServiceTests {
         Application finalApplication = application;
         Mono<Tuple4<Application, List<NewPage>, List<NewAction>, List<ActionCollection>>> importedApplication =
                 applicationJson
-                        .flatMap(applicationJson1 -> importExportApplicationService.mergeApplicationJsonWithApplication(
+                        .flatMap(applicationJson1 -> importApplicationService.mergeApplicationJsonWithApplication(
                                 workspaceId, finalApplication.getId(), null, applicationJson1, new ArrayList<>()))
                         .flatMap(application1 -> {
                             Mono<List<NewPage>> pageList = newPageService
@@ -3652,7 +3654,7 @@ public class ImportExportApplicationServiceTests {
         Application finalApplication = application;
         Mono<Tuple4<Application, List<NewPage>, List<NewAction>, List<ActionCollection>>> importedApplication =
                 applicationJson
-                        .flatMap(applicationJson1 -> importExportApplicationService.mergeApplicationJsonWithApplication(
+                        .flatMap(applicationJson1 -> importApplicationService.mergeApplicationJsonWithApplication(
                                 workspaceId, finalApplication.getId(), "master", applicationJson1, new ArrayList<>()))
                         .flatMap(application1 -> {
                             Mono<List<NewPage>> pageList = newPageService
@@ -3763,7 +3765,7 @@ public class ImportExportApplicationServiceTests {
         Application finalApplication = application;
         Mono<Tuple4<Application, List<NewPage>, List<NewAction>, List<ActionCollection>>> importedApplication =
                 applicationJson
-                        .flatMap(applicationJson1 -> importExportApplicationService.mergeApplicationJsonWithApplication(
+                        .flatMap(applicationJson1 -> importApplicationService.mergeApplicationJsonWithApplication(
                                 workspaceId, branchApp.getId(), "feature", applicationJson1, new ArrayList<>()))
                         .flatMap(application2 -> {
                             Mono<List<NewPage>> pageList = newPageService
@@ -3873,7 +3875,7 @@ public class ImportExportApplicationServiceTests {
         Application finalApplication = application;
         Mono<Tuple4<Application, List<NewPage>, List<NewAction>, List<ActionCollection>>> importedApplication =
                 applicationJson
-                        .flatMap(applicationJson1 -> importExportApplicationService.mergeApplicationJsonWithApplication(
+                        .flatMap(applicationJson1 -> importApplicationService.mergeApplicationJsonWithApplication(
                                 workspaceId, branchApp.getId(), "feature", applicationJson1, List.of("Page1")))
                         .flatMap(application2 -> {
                             Mono<List<NewPage>> pageList = newPageService
@@ -3983,7 +3985,7 @@ public class ImportExportApplicationServiceTests {
         Application finalApplication = application;
         Mono<Tuple4<Application, List<NewPage>, List<NewAction>, List<ActionCollection>>> importedApplication =
                 applicationJson
-                        .flatMap(applicationJson1 -> importExportApplicationService.mergeApplicationJsonWithApplication(
+                        .flatMap(applicationJson1 -> importApplicationService.mergeApplicationJsonWithApplication(
                                 workspaceId, branchApp.getId(), "feature", applicationJson1, List.of("Page1", "Page2")))
                         .flatMap(application2 -> {
                             Mono<List<NewPage>> pageList = newPageService
@@ -4056,7 +4058,7 @@ public class ImportExportApplicationServiceTests {
         Application finalApplication = application;
         Mono<Tuple4<Application, List<NewPage>, List<NewAction>, List<ActionCollection>>> importedApplication =
                 applicationJson
-                        .flatMap(applicationJson1 -> importExportApplicationService.mergeApplicationJsonWithApplication(
+                        .flatMap(applicationJson1 -> importApplicationService.mergeApplicationJsonWithApplication(
                                 workspaceId, finalApplication.getId(), null, applicationJson1, List.of("Page1")))
                         .flatMap(application1 -> {
                             Mono<List<NewPage>> pageList = newPageService
@@ -4131,7 +4133,7 @@ public class ImportExportApplicationServiceTests {
         Application finalApplication = application;
         Mono<Tuple4<Application, List<NewPage>, List<NewAction>, List<ActionCollection>>> importedApplication =
                 applicationJson
-                        .flatMap(applicationJson1 -> importExportApplicationService.mergeApplicationJsonWithApplication(
+                        .flatMap(applicationJson1 -> importApplicationService.mergeApplicationJsonWithApplication(
                                 workspaceId,
                                 finalApplication.getId(),
                                 null,
@@ -4205,7 +4207,7 @@ public class ImportExportApplicationServiceTests {
                 .block();
 
         Mono<ApplicationImportDTO> resultMono =
-                importExportApplicationService.extractFileAndSaveApplication(workspaceId, filePart);
+                importApplicationService.extractFileAndSaveApplication(workspaceId, filePart);
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -4301,7 +4303,7 @@ public class ImportExportApplicationServiceTests {
 
                     return layoutActionService
                             .createSingleAction(action, Boolean.FALSE)
-                            .then(importExportApplicationService.exportApplicationById(
+                            .then(exportApplicationService.exportApplicationById(
                                     objects.getT1().getId(), ""));
                 });
 
@@ -4330,7 +4332,7 @@ public class ImportExportApplicationServiceTests {
                 .block();
 
         Mono<ApplicationJson> resultMono =
-                importExportApplicationService.exportApplicationById(createdApplication.getId(), "");
+                exportApplicationService.exportApplicationById(createdApplication.getId(), "");
 
         StepVerifier.create(resultMono)
                 .assertNext(applicationJson -> {
@@ -4367,7 +4369,7 @@ public class ImportExportApplicationServiceTests {
         PageDTO applicationPageDTO = applicationPageService.createPage(pageDTO).block();
 
         Mono<ApplicationJson> resultMono =
-                importExportApplicationService.exportApplicationById(applicationPageDTO.getApplicationId(), "");
+                exportApplicationService.exportApplicationById(applicationPageDTO.getApplicationId(), "");
 
         StepVerifier.create(resultMono)
                 .assertNext(applicationJson -> {
@@ -4401,7 +4403,7 @@ public class ImportExportApplicationServiceTests {
                             return applicationPageService.createPage(pageDTO).thenReturn(createdApp);
                         })
                         .zipWith(applicationJson)
-                        .flatMap(objects -> importExportApplicationService
+                        .flatMap(objects -> importApplicationService
                                 .restoreSnapshot(
                                         workspaceId,
                                         objects.getT2(),
@@ -4482,7 +4484,7 @@ public class ImportExportApplicationServiceTests {
 
         FilePart filePart = createFilePart("test_assets/ImportExportServiceTest/valid-application.json");
         String workspaceId = createTemplateWorkspace().getId();
-        final Mono<Application> resultMonoWithoutDiscardOperation = importExportApplicationService
+        final Mono<Application> resultMonoWithoutDiscardOperation = importApplicationService
                 .extractFileAndSaveApplication(workspaceId, filePart)
                 .flatMap(applicationImportDTO -> {
                     PageDTO page = new PageDTO();
@@ -4537,7 +4539,7 @@ public class ImportExportApplicationServiceTests {
         // Import the same application again to find if the added page is deleted
         final Mono<Application> resultMonoWithDiscardOperation = resultMonoWithoutDiscardOperation
                 .flatMap(importedApplication -> applicationService.save(importedApplication))
-                .flatMap(savedApplication -> importExportApplicationService.extractFileAndSaveApplication(
+                .flatMap(savedApplication -> importApplicationService.extractFileAndSaveApplication(
                         workspaceId, filePart, savedApplication.getId()))
                 .map(ApplicationImportDTO::getApplication);
 
@@ -4591,8 +4593,8 @@ public class ImportExportApplicationServiceTests {
                 .block();
 
         FilePart filePart = createFilePart("test_assets/ImportExportServiceTest/valid-application.json");
-        final Mono<ApplicationImportDTO> resultMono = importExportApplicationService.extractFileAndSaveApplication(
-                workspaceId, filePart, application.getId());
+        final Mono<ApplicationImportDTO> resultMono =
+                importApplicationService.extractFileAndSaveApplication(workspaceId, filePart, application.getId());
 
         StepVerifier.create(resultMono)
                 .expectErrorMatches(throwable -> throwable instanceof AppsmithException
@@ -4619,7 +4621,7 @@ public class ImportExportApplicationServiceTests {
                 })
                 .cache();
         Mono<ApplicationJson> getExportedAppMono =
-                addJSLibMonoCached.then(importExportApplicationService.exportApplicationById(testAppId, ""));
+                addJSLibMonoCached.then(exportApplicationService.exportApplicationById(testAppId, ""));
         StepVerifier.create(Mono.zip(addJSLibMonoCached, getExportedAppMono))
                 .assertNext(tuple2 -> {
                     Boolean isJSLibAdded = tuple2.getT1();
@@ -4677,8 +4679,8 @@ public class ImportExportApplicationServiceTests {
                 .block();
 
         FilePart filePart = createFilePart("test_assets/ImportExportServiceTest/valid-application.json");
-        final Mono<ApplicationImportDTO> resultMono = importExportApplicationService.extractFileAndSaveApplication(
-                workspaceId, filePart, application.getId());
+        final Mono<ApplicationImportDTO> resultMono =
+                importApplicationService.extractFileAndSaveApplication(workspaceId, filePart, application.getId());
 
         StepVerifier.create(resultMono)
                 .assertNext(applicationImportDTO -> {
@@ -4711,11 +4713,10 @@ public class ImportExportApplicationServiceTests {
                 })
                 .flatMap(application -> {
                     FilePart filePart = createFilePart("test_assets/ImportExportServiceTest/valid-application.json");
-                    return importExportApplicationService
+                    return importApplicationService
                             .extractApplicationJson(filePart)
-                            .flatMap(applicationJson ->
-                                    importExportApplicationService.mergeApplicationJsonWithApplication(
-                                            workspaceId, application.getId(), null, applicationJson, null));
+                            .flatMap(applicationJson -> importApplicationService.mergeApplicationJsonWithApplication(
+                                    workspaceId, application.getId(), null, applicationJson, null));
                 });
 
         StepVerifier.create(applicationImportDTOMono)
@@ -4745,7 +4746,7 @@ public class ImportExportApplicationServiceTests {
                 })
                 .flatMap(application -> {
                     FilePart filePart = createFilePart("test_assets/ImportExportServiceTest/valid-application.json");
-                    return importExportApplicationService.extractFileAndSaveApplication(
+                    return importApplicationService.extractFileAndSaveApplication(
                             workspaceId, filePart, application.getId());
                 });
 
@@ -4841,7 +4842,7 @@ public class ImportExportApplicationServiceTests {
                     return newPageService
                             .updatePage(applicationPage.getId(), pageDTO)
                             // export the application
-                            .then(importExportApplicationService.exportApplicationById(
+                            .then(exportApplicationService.exportApplicationById(
                                     application.getId(), SerialiseApplicationObjective.VERSION_CONTROL));
                 });
 
@@ -4941,7 +4942,7 @@ public class ImportExportApplicationServiceTests {
                     datasource.setName("DS_FOR_RENAME_TEST_RENAMED");
                     return datasourceService
                             .save(datasource)
-                            .then(importExportApplicationService.exportApplicationById(
+                            .then(exportApplicationService.exportApplicationById(
                                     application.getId(), SerialiseApplicationObjective.VERSION_CONTROL));
                 });
